fix: moved eda_process.py to agents folder. Created notebook with the same code in eda_process for cell based running
