+++ services/data/reports/03_numeric_describe.csv
@@ -1,19 +1,19 @@
 column,count,mean,std,min,25%,50%,75%,max
 year,100000.0,2018.36082,1.3452386495600857,2015.0,2019.0,2019.0,2019.0,2022.0
-age,100000.0,41.885856000000004,22.516839871617023,0.08,24.0,43.0,60.0,80.0
-race:AfricanAmerican,100000.0,0.20223,0.4016648359595404,0.0,0.0,0.0,0.0,1.0
+age,100000.0,41.885856000000004,22.51683987161702,0.08,24.0,43.0,60.0,80.0
+race:AfricanAmerican,100000.0,0.20223,0.40166483595954033,0.0,0.0,0.0,0.0,1.0
 race:Asian,100000.0,0.20015,0.40011445664432593,0.0,0.0,0.0,0.0,1.0
 race:Caucasian,100000.0,0.19876,0.3990689852150247,0.0,0.0,0.0,0.0,1.0
 race:Hispanic,100000.0,0.19888,0.3991595406393148,0.0,0.0,0.0,0.0,1.0
 race:Other,100000.0,0.19998,0.3999869991587163,0.0,0.0,0.0,0.0,1.0
-hypertension,100000.0,0.07485,0.26315047022891647,0.0,0.0,0.0,0.0,1.0
+hypertension,100000.0,0.07485,0.2631504702289164,0.0,0.0,0.0,0.0,1.0
 heart_disease,100000.0,0.03942,0.1945930169980995,0.0,0.0,0.0,0.0,1.0
-bmi,100000.0,27.3207671,6.636783416648369,10.01,23.63,27.32,29.58,95.69
+bmi,100000.0,27.320767099999994,6.636783416648368,10.01,23.63,27.32,29.58,95.69
 hbA1c_level,100000.0,5.527507,1.0706720918835437,3.5,4.8,5.8,6.2,9.0
 blood_glucose_level,100000.0,138.05806,40.70813604870415,80.0,100.0,140.0,159.0,300.0
-sleep_hours,100000.0,6.49388,1.7083300996930586,4.0,5.0,6.0,8.0,9.0
-family_history,100000.0,0.30013,0.4583165757866896,0.0,0.0,0.0,1.0,1.0
-medication_use,100000.0,0.15045,0.357513741918764,0.0,0.0,0.0,0.0,1.0
-gestational_history,58552.0,0.09933050963246345,0.2991078191137739,0.0,0.0,0.0,0.0,1.0
+sleep_hours,100000.0,6.49388,1.7083300996930584,4.0,5.0,6.0,8.0,9.0
+family_history,100000.0,0.30013,0.45831657578668966,0.0,0.0,0.0,1.0,1.0
+medication_use,100000.0,0.15045,0.35751374191876406,0.0,0.0,0.0,0.0,1.0
+gestational_history,58552.0,0.09933050963246345,0.29910781911377393,0.0,0.0,0.0,0.0,1.0
 environmental_risk,100000.0,5.50657,2.8746859782060947,1.0,3.0,6.0,8.0,10.0
 diabetes,100000.0,0.085,0.27888308976662174,0.0,0.0,0.0,0.0,1.0
